second commit

+++ src/org/foodie/dao/ProductDao.java
@@ -41,7 +41,7 @@ public class ProductDao {
         ps.setString(1, getNewProductId());
         ps.setString(2, product.getCompanyId());
         ps.setString(3, product.getProductName());
-        ps.setDouble(4, product.getProductPrice());
+        ps.setInt(4, product.getProductPrice());
         ps.setBinaryStream(5, new ByteArrayInputStream(imageData), imageData.length);
         int x = ps.executeUpdate();
         return x > 0;
@@ -59,7 +59,7 @@ public class ProductDao {
             product.setCompanyId(rs.getString("company_id"));
             product.setProductId(rs.getString("product_id"));
             product.setProductName(rs.getString("product_name"));
-            product.setProductPrice(rs.getDouble("product_price"));
+            product.setProductPrice(rs.getInt("product_price"));
 
             //convert the blob to image : 
             // Fetch the image data as InputStream
@@ -97,7 +97,7 @@ public class ProductDao {
             product.setCompanyId(rs.getString("company_id"));
             product.setProductId(rs.getString("product_id"));
             product.setProductName(rs.getString("product_name"));
-            product.setProductPrice(rs.getDouble("product_price"));
+            product.setProductPrice(rs.getInt("product_price"));
             //convert the blob to image : 
             // Fetch the image data as InputStream
             InputStream inputStream = rs.getBinaryStream("product_image");
@@ -126,7 +126,7 @@ public class ProductDao {
             product.setCompanyId(rs.getString("company_id"));
             product.setProductId(rs.getString("product_id"));
             product.setProductName(rs.getString("product_name"));
-            product.setProductPrice(rs.getDouble("product_price"));
+            product.setProductPrice(rs.getInt("product_price"));
             //convert the blob to image : 
             // Fetch the image data as InputStream
             InputStream inputStream = rs.getBinaryStream("product_image");
